backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.2.0
python-dotenv==1.0.1
gunicorn==21.2.0
# Database dependencies
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
# Structured logging
structlog==23.2.0
# Authentication dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
python-multipart==0.0.9